+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 0.3.0
+- Show standings page teams as pitch view and with gameweeks separately
+- Setup transfer page with correct handling of which gameweek to set
+- Updating icons in ribbon
+
 ### 0.2.0
 - Created gameweek tables in SQL & add to automatic setup
 - Implemented Draft picking logic